+++ zcash_client_sqlite/src/wallet/init/migrations/addresses_table.rs
@@ -90,9 +90,9 @@ impl<P: consensus::Parameters> RusqliteMigration for Migration<P> {
                     "Address in accounts table was not a Unified Address.".to_string(),
                 ));
             };
-            let (expected_address, idx) = ufvk.default_address(Some(
-                UnifiedAddressRequest::unsafe_new(Omit, Require, UA_TRANSPARENT),
-            ))?;
+            let (expected_address, idx) = ufvk.default_address(
+                UnifiedAddressRequest::unsafe_custom(Omit, Require, UA_TRANSPARENT),
+            )?;
             if decoded_address != expected_address {
                 return Err(WalletMigrationError::CorruptedData(format!(
                     "Decoded UA {} does not match the UFVK's default address {} at {:?}.",
@@ -162,8 +162,10 @@ impl<P: consensus::Parameters> RusqliteMigration for Migration<P> {
                 ],
             )?;
 
-            let (address, d_idx) = ufvk.default_address(Some(
-                UnifiedAddressRequest::unsafe_new(Omit, Require, UA_TRANSPARENT),
+            let (address, d_idx) = ufvk.default_address(UnifiedAddressRequest::unsafe_custom(
+                Omit,
+                Require,
+                UA_TRANSPARENT,
             ))?;
             insert_address(transaction, &self.params, account, d_idx, &address)?;
         }