+++ zcash_keys/CHANGELOG.md
@@ -7,7 +7,22 @@ and this library adheres to Rust's notion of
 ## [Unreleased]
 
 ### Added
-- `zcash_keys::keys::UnifiedAddressRequest::{sapling, orchard, p2pkh}` methods
+- `zcash_keys::keys::UnifiedAddressRequest::{sapling, orchard, p2pkh}`
+- `zcash_keys::keys::UnifiedIncomingViewingKey::{has_sapling, has_orchard, has_p2pkh,
+  receiver_requirements, to_receiver_requirements}`
+- `zcash_keys::keys::ReceiverRequirements`
+
+### Changed
+- `zcash_keys::keys::UnifiedAddressRequest` is now an enum instead of a struct.
+  The `new` and `unsafe_new` methods have been replaced by `custom` and
+  `unsafe_custom` respectively.
+
+### Removed
+- `UnifiedAddressRequest::{new, unsafe_new}`: use `{custom, unsafe_custom}`
+  respectively instead.
+- `UnifiedAddressRequest::{intersect, orchard, sapling, p2pkh}`: replacements
+  for these methods are now provided on the newly-added `ReceiverRequirements`
+  type.
 
 ## [0.7.0] - 2025-02-21